pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "security-gym"
version = "0.1.0"
description = "Gymnasium environments for cybersecurity threat detection with continual learning"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [{ name = "Keith Lawson" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
]
keywords = ["gymnasium", "reinforcement-learning", "cybersecurity", "continual-learning"]
dependencies = [
    "gymnasium>=1.0.0",
    "numpy>=1.24.0",
    "mmh3>=4.0.0",
]

[project.optional-dependencies]
alberta = ["alberta-framework>=0.8.0", "jax>=0.4.20", "jaxlib>=0.4.20"]
collection = ["watchdog>=3.0.0"]
attacks = ["paramiko>=3.0.0", "requests>=2.28.0", "pyyaml>=6.0", "scapy>=2.5.0"]
dev = ["pytest>=7.0", "ruff>=0.4.0"]
all = ["security-gym[alberta,collection,attacks,dev]"]

[project.urls]
Homepage = "https://github.com/j-klawson/security-gym"
Documentation = "https://j-klawson.github.io/security-gym"
Repository = "https://github.com/j-klawson/security-gym"

[project.entry-points."gymnasium.envs"]
__root__ = "security_gym.envs:register_envs"

[tool.hatch.build.targets.wheel]
packages = ["src/security_gym"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["security_gym.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["scapy.*"]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests"]
